Split out path key api into separate module
